+++ blocks/header/header.js
@@ -11,7 +11,7 @@ import { loadFragment } from '../fragment/fragment.js';
 
 import renderAuthCombine from './renderAuthCombine.js';
 import { renderAuthDropdown } from './renderAuthDropdown.js';
-import { applyHashTagsForDomElement } from '../../scripts/hashTags.js';
+import { applyHashTagsForDomElement } from '../../scripts/api/hashtags/api.js';
 
 // media query match that indicates mobile/tablet width
 const isDesktop = window.matchMedia('(min-width: 900px)');
